Mv test_apm_config to subdir
